Adjust event probability
Why?

Two reasons:
1.) "Skimpy plate mail" event is drowned out by all other events with base modifiers of 50+. That's not 50%, that's x50 event probability! So, plate armor event has total weight of 50,  "pardon my bold proposal" has roughly weight 100*50*2=100000, making sure you'll never ever get a skimpy plate armor when you are attractive.

2.) Right now, you gain "Pardon my bold proposal" every-fucking-year if your attraction is 20+ (and you aren't eligible for other events). Like clockwork. Subjective if that's a problem, surely, but I find that weird.

Also, the event should trigger more often with high attraction - yes, I can see the weight modifier is already there, but the way it is set up that's only relevant to drown out to compare to other events, not to a "null" event. Basically, if you are very attractive you gain more "bold proposals", less "sub/dom" content - but you always gain 1 event regardless, so if you only qualify for bold proposal, you always gain that one.

Now one *could* just reduce the "chance_to_happen = 100" in LIDA_random_events, but I think it's more useful to have a "nothing is happening event" so that IF you are eligible for multiple events, your chance to trigger one of them this year increases. The weights are altered with that in mind, about 50/50 chance for a proposal each year with 30 attraction.